[RENAME] 쿠폰 이벤트 파일 디렉토리 이동
