+++ src/main/java/kr/hhplus/be/server/application/coupon/CouponEventService.java
@@ -2,9 +2,9 @@ package kr.hhplus.be.server.application.coupon;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import kr.hhplus.be.server.interfaces.web.coupon.dto.CouponIssuedEventDto;
-import kr.hhplus.be.server.infrastructure.external.coupon.CouponOutBoxEvent;
-import kr.hhplus.be.server.infrastructure.external.coupon.CouponOutBoxEventRepository;
+import kr.hhplus.be.server.interfaces.web.coupon.dto.event.CouponIssuedEventDto;
+import kr.hhplus.be.server.domain.coupon.event.CouponOutBoxEvent;
+import kr.hhplus.be.server.domain.coupon.event.CouponOutBoxEventRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
