+++ src/main/java/kr/hhplus/be/server/domain/coupon/event/CouponOutBoxEventRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infrastructure.external.coupon;
+package kr.hhplus.be.server.domain.coupon.event;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
