+++ src/main/java/kr/hhplus/be/server/domain/order/event/OrderCompletedEventListener.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.domain.order.event;
 
-import kr.hhplus.be.server.infrastructure.external.payment.DataPlatform;
+import kr.hhplus.be.server.infrastructure.external.orderinfo.DataPlatform;
 import kr.hhplus.be.server.interfaces.web.order.dto.event.OrderCompletedEventDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
