+++ src/main/java/kr/hhplus/be/server/infrastructure/external/orderinfo/DataPlatform.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.infrastructure.external.payment;
+package kr.hhplus.be.server.infrastructure.external.orderinfo;
 
 import kr.hhplus.be.server.interfaces.web.order.dto.event.OrderCompletedEventDto;
 import lombok.extern.slf4j.Slf4j;
